Node.js MySql How to Verify JWT (Authorization)

+++ api/controllers/post.js
@@ -1,4 +1,5 @@
 import { db } from '../db.js';
+import jwt from 'jsonwebtoken';
 
 export const getPosts = (req, res, next) => {
   const q = req.query.cat
@@ -6,7 +7,7 @@ export const getPosts = (req, res, next) => {
     : 'SELECT * FROM posts';
 
   db.query(q, [req.query.cat], (err, data) => {
-    if (err) return res.json(err);
+    if (err) return res.status(500).json(err);
     return res.status(200).json(data);
   });
 };
@@ -16,7 +17,7 @@ export const getPost = (req, res, next) => {
     'SELECT users.username, users.img AS userImg, posts.title, posts.description, posts.img, posts.cat, posts.date FROM posts JOIN users ON users.id = posts.uid WHERE posts.id = ?';
 
   db.query(q, [req.params.id], (err, data) => {
-    if (err) return res.json(err);
+    if (err) return res.status(500).json(err);
     return res.status(200).json(data[0]);
   });
 };
@@ -24,11 +25,20 @@ export const getPost = (req, res, next) => {
 export const addPost = (req, res, next) => {};
 
 export const deletePost = (req, res, next) => {
-  const q = 'DELETE FROM posts WHERE id = ?';
+  const token = req.cookies.access_token;
 
-  db.query(q, [req.params.id], (err, data) => {
-    if (err) return res.json(err);
-    return res.status(200).json('The post has been deleted!');
+  if (!token) return res.status(401).json('Not authenticated!');
+
+  jwt.verify(token, 'jwtkey', (err, userInfo) => {
+    if (err) return res.status(403).json('Token is not valid!');
+
+    const postId = req.params.id;
+    const q = 'DELETE FROM posts WHERE id = ? AND uid = ?';
+
+    db.query(q, [postId, userInfo.id], (err, data) => {
+      if (err) return res.status(403).json('You can delete only your post!');
+      return res.status(200).json('Post has been deleted!');
+    });
   });
 };
